fix: handle visual highlight reset for `vs` and `Vs` mode
In e284230, we remove the "Listen all the time for `<Esc>` key", and
this caused an issue with `v_CTRL-O` which kept the highlight in visual
mode after returning to select mode.

This change fix the issue by changing the pattern, the pattern means
"from every visual mode to any other mode except visual".

See: `:h v_CTRL-O`